+++ lib/src/filter_list_widget.dart
@@ -72,6 +72,7 @@ class FilterListWidget<T extends Object> extends StatelessWidget {
     this.hideHeader = false,
     this.backgroundColor = Colors.white,
     this.enableOnlySingleSelection = false,
+    this.maximumSelectionLength,
     this.allButtonText = 'All',
     this.applyButtonText = 'Apply',
     this.resetButtonText = 'Reset',
@@ -119,6 +120,12 @@ class FilterListWidget<T extends Object> extends StatelessWidget {
   ///
   /// Default value is `false`
   final bool enableOnlySingleSelection;
+
+  /// if `maximumSelectionLength` is not null then it will limit the maximum selection length.
+  /// `maximumSelectionLength` should be greater than 0. If `maximumSelectionLength` is less than 0 then it will throw an exception.
+  /// Only works when `enableOnlySingleSelection` is false.
+  /// Default value is [null]
+  final int? maximumSelectionLength;
 
   /// The `onApplyButtonClick` is a callback which return list of all selected items on apply button click.  if no item is selected then it will return empty list.
   final OnApplyButtonClick<T>? onApplyButtonClick;
@@ -180,9 +187,27 @@ class FilterListWidget<T extends Object> extends StatelessWidget {
                   hideCloseIcon: hideCloseIcon,
                   headerCloseIcon: headerCloseIcon,
                   onSearch: (String value) {
+                    final stateList = FilterState.of<T>(context).items;
+
+                    // Reset filter list if search box is empty
                     if (value.isEmpty) {
                       FilterState.of<T>(context).items = listData;
                       return;
+                    }
+                    // Reassign items to filter list when it is empty but local list has data
+                    else if (stateList != null &&
+                        stateList.isEmpty &&
+                        listData != null &&
+                        listData!.isNotEmpty) {
+                      final isFoundInLocalState =
+                          listData!.any((item) => onItemSearch(item, value));
+
+                      if (isFoundInLocalState) {
+                        FilterState.of<T>(context).items = listData!
+                            .where((item) => onItemSearch(item, value))
+                            .toList();
+                        return;
+                      }
                     }
                     FilterState.of<T>(context)
                         .filter((item) => onItemSearch(item, value));
@@ -206,6 +231,7 @@ class FilterListWidget<T extends Object> extends StatelessWidget {
                   enableOnlySingleSelection: enableOnlySingleSelection,
                   validateSelectedItem: validateSelectedItem,
                   validateRemoveItem: validateRemoveItem,
+                  maximumSelectionLength: maximumSelectionLength,
                 ),
               ),
             ],
@@ -214,6 +240,7 @@ class FilterListWidget<T extends Object> extends StatelessWidget {
           // /// Bottom section for control buttons
           ControlButtonBar<T>(
             controlButtons: controlButtons,
+            maximumSelectionLength: maximumSelectionLength,
             allButtonText: allButtonText,
             applyButtonText: applyButtonText,
             resetButtonText: resetButtonText,