+++ lib/src/widget/choice_list.dart
@@ -12,12 +12,14 @@ class ChoiceList<T> extends StatelessWidget {
     this.choiceChipLabel,
     this.enableOnlySingleSelection = false,
     this.validateRemoveItem,
+    this.maximumSelectionLength,
   }) : super(key: key);
   final ValidateSelectedItem<T> validateSelectedItem;
   final ChoiceChipBuilder? choiceChipBuilder;
   final LabelDelegate<T>? choiceChipLabel;
   final bool enableOnlySingleSelection;
   final ValidateRemoveItem<T>? validateRemoveItem;
+  final int? maximumSelectionLength;
 
   List<Widget> _buildChoiceList(BuildContext context) {
     final theme = FilterListTheme.of(context).controlBarButtonTheme;
@@ -30,9 +32,15 @@ class ChoiceList<T> extends StatelessWidget {
     final List<Widget> choices = [];
     for (final item in items) {
       final selected = validateSelectedItem(selectedListData, item);
+
+      // Check if maximum selection length reached
+      final maxSelectionReached = maximumSelectionLength != null &&
+          state.selectedItems != null &&
+          state.selectedItems!.length >= maximumSelectionLength!;
       choices.add(
         ChoiceChipWidget(
           choiceChipBuilder: choiceChipBuilder,
+          disabled: maxSelectionReached,
           item: item,
           onSelected: (value) {
             if (enableOnlySingleSelection) {
@@ -48,6 +56,10 @@ class ChoiceList<T> extends StatelessWidget {
                   state.removeSelectedItem(item);
                 }
               } else {
+                // Add maximum selection length check
+                if (maxSelectionReached && !selected) {
+                  return;
+                }
                 state.addSelectedItem(item);
               }
             }